Fixed test case on encryption

+++ core/src/test/java/com/orientechnologies/orient/core/encryption/impl/ODESEncryptionTest.java
@@ -154,6 +154,7 @@ public class ODESEncryptionTest extends AbstractEncryptionTest {
       } catch (OSecurityException e) {
         Assert.assertTrue(true);
       } finally {
+        db.activateOnCurrentThread();
         db.close();
         Orient.instance().getStorage(DBNAME_CLUSTERTEST).close(true, false);
       }
@@ -166,6 +167,7 @@ public class ODESEncryptionTest extends AbstractEncryptionTest {
       } catch (OSecurityException e) {
         Assert.assertTrue(true);
       } finally {
+        db.activateOnCurrentThread();
         db.close();
         Orient.instance().getStorage(DBNAME_CLUSTERTEST).close(true, false);
       }
@@ -177,7 +179,8 @@ public class ODESEncryptionTest extends AbstractEncryptionTest {
 
     } finally {
       db.activateOnCurrentThread();
-      db.drop();
+      if (db.exists())
+        db.drop();
     }
   }
 }
